Remove unnecesary and buggy name parameter
This is the namevar, so it will default to the title of the resource in most
cases. However, when $package_name is an array, it will create a separate
resource for each element, while still passing in the entire array. Supplying an
array for the name paramater was deprecated in Puppet 3.4.0, causing the catalog
to fail when run on an Ubuntu system.